Refactoring code and almost ready for tests.
A refactor to make a similar output of the PBI with other exact indexes.
We can try the implementation comparing the output of the PBI against
AESA or another index, with that we can know the presicion of our
PBI implementation. Folders with data, queries and built executables
are not included in the repo.